backend/requirements.txt
asgiref==3.10.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.4
contourpy==1.3.1
cycler==0.12.1
dataclasses==0.6
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.2.31
dnspython==2.8.0
fonttools==4.55.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.15.3
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.2
requests==2.32.5
seaborn==0.13.2
selenium==4.34.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
